Added Selfbot Command Checks
I've added the selfbot command checks.

basically if a command has selfbot: true in its exports.conf

On boot it will skip the command and output in console this https://u.pomf.is/aywkdk.png

On command reload... it just fails silently but the bot continues running

On download it will output the same as in the console but in a message instead.