Fixed pieces count
Fixed an issue with the counting of pieces (separating them according to
code language, that is) where function counts (possibly other types of
pieces, as well) would include some things that aren't functions (that
happen to also be stored in `client.funcs`).

Also, pieces counts are accurate now, since all promises are awaited.